chore(agenda): move static methods to AgendaUtils
use AgendaUtils as a namespace to reduce line file length in Agenda
